Rearrange Secret
