+++ materialize/datasources/datasource_secret.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"database/sql"
 	"errors"
-	"fmt"
 	"log"
-	"strings"
+
+	"terraform-materialize/materialize/materialize"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -15,7 +15,7 @@ import (
 
 func Secret() *schema.Resource {
 	return &schema.Resource{
-		ReadContext: secretRead,
+		ReadContext: SecretRead,
 		Schema: map[string]*schema.Schema{
 			"database_name": {
 				Type:        schema.TypeString,
@@ -57,41 +57,14 @@ func Secret() *schema.Resource {
 	}
 }
 
-func secretQuery(databaseName, schemaName string) string {
-	q := strings.Builder{}
-	q.WriteString(`
-		SELECT
-			mz_secrets.id,
-			mz_secrets.name,
-			mz_schemas.name,
-			mz_databases.name
-		FROM mz_secrets
-		JOIN mz_schemas
-			ON mz_secrets.schema_id = mz_schemas.id
-		JOIN mz_databases
-			ON mz_schemas.database_id = mz_databases.id`)
-
-	if databaseName != "" {
-		q.WriteString(fmt.Sprintf(`
-		WHERE mz_databases.name = '%s'`, databaseName))
-
-		if schemaName != "" {
-			q.WriteString(fmt.Sprintf(` AND mz_schemas.name = '%s'`, schemaName))
-		}
-	}
-
-	q.WriteString(`;`)
-	return q.String()
-}
-
-func secretRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+func SecretRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
 
 	conn := meta.(*sqlx.DB)
 
 	schemaName := d.Get("schema_name").(string)
 	databaseName := d.Get("database_name").(string)
-	q := secretQuery(databaseName, schemaName)
+	q := materialize.ReadSecretDatasource(databaseName, schemaName)
 
 	rows, err := conn.Query(q)
 	if errors.Is(err, sql.ErrNoRows) {
@@ -123,15 +96,6 @@ func secretRead(ctx context.Context, d *schema.ResourceData, meta interface{}) d
 		return diag.FromErr(err)
 	}
 
-	if databaseName != "" && schemaName != "" {
-		id := fmt.Sprintf("%s|%s|secrets", databaseName, schemaName)
-		d.SetId(id)
-	} else if databaseName != "" {
-		id := fmt.Sprintf("%s|secrets", databaseName)
-		d.SetId(id)
-	} else {
-		d.SetId("secrets")
-	}
-
+	SetId("secrets", databaseName, schemaName, d)
 	return diags
 }
